+++ src/js/view/room/modal/library.mjs
@@ -602,9 +602,10 @@ function assetToPreview (asset) {
   node.dataset.a = asset.id
 
   if (piece._meta.hasColor) {
-    piece.c[0] = Number.parseInt(asset.bg)
+    const colors = getTemplate().colors
+    piece.c[0] = Number.parseInt(asset.bg) % colors.length
     if (piece.c[0] !== 0) {
-      node.css({ '--fbg-color': getTemplate().colors[piece.c[0] - 1].value })
+      node.css({ '--fbg-color': colors[piece.c[0] - 1].value })
     }
   }